pyproject.toml
[project]
name = "crashinette"
version = "0.1.0"
description = "Destroy your minishell"
authors = [
    {name = "Leonardo Lopes Pereira"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pexpect (>=4.9.0,<5.0.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
